+++ crates/bevy_mod_scripting_world/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "bevy_mod_scripting_world"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+keywords.workspace = true
+categories.workspace = true
+readme.workspace = true
+
+[dependencies]
+
+[lints]
+workspace = true
